renamed CardDto to NFCDto, cleanup on link card form and script
